[C] Rename + organize files
