+++ packages/admin/components/forms/EntitySelector/EntitySelectorItem/index.ts
@@ -1,0 +1,3 @@
+import EntitySelectorItem from "./EntitySelectorItem";
+
+export default EntitySelectorItem;
