--- packages/admin/components/forms/EntitySelectorUI/index.ts
@@ -1,3 +1,0 @@
-import EntitySelectorUI from "./EntitySelectorUI";
-
-export default EntitySelectorUI;
